requirements-lambda.txt
# Dependências essenciais para Lambda (versão otimizada)
# Apenas as bibliotecas realmente necessárias para reduzir o tamanho do pacote

# Core dependencies para PDF processing
pymupdf==1.26.3
pillow==11.3.0

# AWS SDK assíncrono (necessário para o código atual)
aioboto3==15.0.0
aiobotocore==2.23.0
boto3==1.38.27
botocore==1.38.27

# Logging estruturado
structlog==25.4.0

# ULID para nomes únicos
ulid-py==1.1.0

# Pydantic para configuração (versão mais leve)
pydantic==2.11.7
pydantic-core==2.33.2

# Dependências de tipo (necessárias para Python < 3.13)
typing_extensions==4.14.1

# Dependências adicionais necessárias
annotated-types==0.7.0
python-dotenv==1.1.1

# Dependências mínimas do PyMuPDF
# (PyMuPDF já inclui suas dependências principais)
